changed Float to Double

+++ src/main/kotlin/rmc/dto/rental/RentalDTO.kt
@@ -16,12 +16,12 @@ data class RentalDTO(
     val userId: Int,
     @Serializable(with = LocalDateSerializer::class)
     val date: LocalDate,
-    val price: Float,
+    val price: Double,
     @Serializable(with = BigDecimalSerializer::class)
     val latitude: BigDecimal,
     @Serializable(with = BigDecimalSerializer::class)
     val longitude: BigDecimal,
     val status: RentalStatus,
-    val distanceTravelled: Float,
+    val distanceTravelled: Double,
     val score: Int
 )
